refactor: Вынесет код работы с БД в репозиторий
