+++ src/modules/wb-parser/types/spreadsheet.type.ts
