+++ src/modules/wb-parser/wb-tariffs-parser.service.ts
@@ -4,19 +4,12 @@ import * as google from "googleapis";
 import { AxiosInstance, AxiosResponse } from "axios";
 
 import { createAPI } from "#services/api.js";
-import knex from "#postgres/knex.js";
 
-import { CRON_SCHEDULE, KNEX_BATCH_CHUNK_SIZE } from "./constants.js";
+import { CRON_SCHEDULE } from "./constants.js";
 
-import {
-    WarehouseType,
-    WBAdaptedDataType,
-    WBDataType,
-} from "./types/wb-data.type.js";
-import { SpreadsheetDataRowType } from "./types/spreadsheed.type.js";
+import { WBDataType } from "./types/wb-data.type.js";
 import { formatDate } from "#utils/date.js";
-import { splitToChunks } from "#utils/common.js";
-
+import { WBTariffsParserRepository } from "./wb-tariffs-parser.repository.js";
 export class WBTariffsParser {
     private readonly loggerPrefix = "[WBParser]";
     private readonly apiBaseUrl =
@@ -24,11 +17,13 @@ export class WBTariffsParser {
     private readonly wbTariffsBoxUri = "box";
     private readonly apiKey = process.env.WB_API_KEY as string;
 
+    private readonly wbTariffsParserRepository: WBTariffsParserRepository;
     private syncTask: ScheduledTask | null = null;
     private api: AxiosInstance;
 
     constructor() {
         this.api = createAPI(this.apiBaseUrl, this.apiKey);
+        this.wbTariffsParserRepository = new WBTariffsParserRepository();
     }
 
     public async init() {
@@ -73,7 +68,7 @@ export class WBTariffsParser {
             }
 
             // Save tariffs to database
-            await this.saveTariffs(wbData);
+            await this.wbTariffsParserRepository.saveTariffs(wbData);
 
             // Insert data to Google Spreadsheet
             // await this.insertDataToGoogleSpreadSheet(wbData);
@@ -110,7 +105,8 @@ export class WBTariffsParser {
     }
 
     private async insertDataToGoogleSpreadSheet(data: WBDataType | null) {
-        const spreadsheetIds = await this.getSpreadsheetIds();
+        const spreadsheetIds =
+            await this.wbTariffsParserRepository.getSpreadsheetIds();
 
         // Auth in Google APIs
         const googleServiceKeysFileName = process.env
@@ -139,102 +135,5 @@ export class WBTariffsParser {
             scopes: ["https://www.googleapis.com/auth/spreadsheets"],
         });
         const sheets = new google.sheets_v4.Sheets({ auth });
-    }
-
-    /*
-        REPOSITORY
-        TODO: Transfer all code below to original repository module
-    */
-    private async getSpreadsheetIds(): Promise<string[] | void> {
-        // Get Spreadsheet IDs
-        const googleSheetRows: SpreadsheetDataRowType[] = await knex
-            .select("*")
-            .from("spreadsheets");
-        const googleSheetIds = googleSheetRows.map((row) => row.spreadsheet_id);
-
-        if (!googleSheetIds || googleSheetIds.length <= 0) {
-            console.error(
-                "Can`t find any Spreadsheet IDs in properly database table",
-            );
-
-            return;
-        }
-
-        console.log(
-            `Found ${googleSheetIds.length} Google Spreadsheet ids to insert in`,
-        );
-
-        return googleSheetIds;
-    }
-
-    private async saveTariffs(data: WBDataType): Promise<void> {
-        const adaptedData = this.adaptWBTariffsToDatabase(data);
-        const savedData =
-            await this.batchInsertTariffsIgnoreOnConflicts(adaptedData);
-
-        const savedDataLength = savedData.length || 0;
-
-        console.log(
-            `Saved box tariffs ${savedDataLength} / ${adaptedData.length}`,
-        );
-    }
-
-    private async batchInsertTariffsIgnoreOnConflicts(
-        adaptedData: WBAdaptedDataType[],
-    ): Promise<WBAdaptedDataType[]> {
-        const chunkedData = splitToChunks(
-            adaptedData as [],
-            KNEX_BATCH_CHUNK_SIZE,
-        );
-        const insertedTariffs: WBAdaptedDataType[] = [];
-
-        await knex.transaction(async (trx) => {
-            for (const chunk of chunkedData) {
-                const savedChunk = await trx<WBAdaptedDataType>(
-                    "wb_tariffs_boxes",
-                )
-                    .insert(chunk)
-                    .onConflict("date_warehouse_index")
-                    .merge()
-                    .returning("*");
-
-                insertedTariffs.push(...(savedChunk as WBAdaptedDataType[]));
-            }
-        });
-
-        return insertedTariffs;
-    }
-
-    private adaptWBTariffsToDatabase(data: WBDataType): WBAdaptedDataType[] {
-        const today = formatDate();
-        const { dtNextBox, dtTillMax, warehouseList } = data.response.data;
-
-        const adaptedData: WBAdaptedDataType[] = warehouseList.map(
-            (warehouse: WarehouseType) => {
-                const {
-                    boxDeliveryAndStorageExpr,
-                    boxDeliveryBase,
-                    boxDeliveryLiter,
-                    boxStorageBase,
-                    boxStorageLiter,
-                    warehouseName,
-                } = warehouse;
-
-                return {
-                    date: today,
-                    dt_next_box: dtNextBox,
-                    dt_till_max: dtTillMax,
-                    box_delivery_and_storage_expr: boxDeliveryAndStorageExpr,
-                    box_delivery_base: boxDeliveryBase,
-                    box_delivery_liter: boxDeliveryLiter,
-                    box_storage_base: boxStorageBase,
-                    box_storage_liter: boxStorageLiter,
-                    warehouse_name: warehouseName,
-                    date_warehouse_index: `${today}|${warehouseName}`,
-                };
-            },
-        );
-
-        return adaptedData;
     }
 }
